simplify a little the implementation for quadratic hex need to test it
git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5706 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ tools/mbcoupler/ElemUtil.hpp
@@ -158,7 +158,7 @@ namespace ElemUtil {
     protected:
       /* Preimages of the vertices -- "canonical vertices" -- are known as "corners".
        * there are 27 vertices for a tri-quadratic xes*/
-      static const double corner[27][3];
+      static const int corner[27][3];
       static const double gauss[8][2];// TODO fix me
       static const unsigned int corner_count = 27;
       static const unsigned int gauss_count  = 8; // TODO fix me
